Redirect HTTP to HTTPS
Automated Certificate Management is now turned on for this Heroku app.
https://blog.heroku.com/announcing-automated-certificate-management

We always get HTTP on our app, though.
TLS is terminated by the Heroku proxy.

http://rss.thoughtbot.com/ and https://rss.thoughtbot.com/
both currently work.

So, let's redirect all HTTP traffic to HTTPS.
This can't be done at the DNS level:
https://support.dnsimple.com/articles/redirect-heroku/#redirect-http2https

It must be done at the application level.

Hardcode the host in the conditional branch to make local testing easy.