Use itunes:subtitle as podcast episode description
Currently, the Giant Robots newsletter includes show notes on every
podcast episode, which can include advertisements and 5-10 links. We
received feedback from our audience that these links are not helpful,
and have decided to remove them.

The weekly Giant Robots newsletter is sent from the Giant Robots RSS
Mailchimp campaign and uses rss.thoughtbot.com as the source of some of
its content. The feed is a consolidation of various other feeds, and
uses an attribute that includes the episode's show notes as the
description. As a result, the episode show notes are also included in
the newsletter.

There was existing work done in August 2019 (not yet deployed) that
modified the RSS feed to use the itunes:summary attribute as the
description for podcast episodes. This attribute, however, still
includes the episode's show notes. Using the itunes:subtitle attribute
instead gives a description without the episode's show notes.